Add files via upload
HTML calculator is used for performing basic mathematical operations like Addition, subtraction, multiplication, and division.To design the basic calculator, we will use HTML, CSS, and JavaScript. HTML is used to design the basic structure of the calculator. CSS styles are used to apply styles on the calculator and JavaScript is used to add the calculation functionality